requirements.txt
allure-pytest==2.13.2
pytest==7.4.0
requests==2.31.0
jsonschema==4.19.0
faker==19.3.0
psycopg2-binary==2.9.7
flake8==6.1.0
flake8-bugbear==23.7.10
flake8-pytest-style==1.7.2
pydantic==2.3.0
pytest-rerunfailures==12.0
pydantic==2.3.0
pyotp==2.9.0
pytest-xdist==3.3.1